improvements to per_sample_log_density documentation
tried to work out how to ensure that no internal shape errors can
occur but failed so far.
in response, clarified the documentation as to what shape is expected
for all random variables occuring in the model (latent or not).

removed note (and code) reflecting doubt about which of the previous two
implementation variants would be more performant