per_sample_log_density adapted to case where no random var has samples
in this case it just behaves as numpyro's log_density and returns
the summed total loss as a scalar.
this can happen during svi if the latent variables are global and thus
the guide does not have a per-sample contribution (but it's effect
gets distributed over all samples).
per_sample_elbo loss fn deals with this case accordingly.